sunxi: Use 128K offset for SPL
Recently it was discovered that sunxi also support for
SPL binary at 128k offset on sdcard or EMMC.
https://groups.google.com/forum/#!topic/linux-sunxi/MaiijyaAFjk

UBoot support was added by the commits:
067e0b9684d4 ("sunxi: Allow booting from 128KB SD/eMMC offset")
7c841d844f ("sunxi: Automate loading from 128KB MMC offset")

We can no longer cut GPT and reuse sdcard tools for another platforms.
But the major benefit that we can use uboot-fastboot to generate
correct partition layout out of the box.

Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>