sunxi: Enable KMS scalling for sunxi devices.
drm_hwcomposer will disable KMS scaling for UI, but will keep it
for video layers.

Should give a benefit playing video on 1080 or higher resolution
displays. As a drawback there is a glitches when video starts to
play and when finish.

Keep disabled for pinephone to avoid dealing with glithes without
real benefits on vertical-oriented displays.

Glitches are caused by bug in kernel driver or display hardware.

Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>